+++ src/models/rate-limit-rules/rate-limit-rule-entity.ts
@@ -3,6 +3,7 @@ export type RateLimitRuleUpdateParams = {
   method?: string;
   tier?: number;
   options?: RateLimitRuleOptions;
+  apiKey?: string;
 };
 
 export type RateLimitRuleOptions = {
@@ -16,6 +17,7 @@ export type RateLimitRuleEntityParams = {
   route: string;
   method: string;
   tier: number;
+  api_key: string;
   options: RateLimitRuleOptions;
   created_at: string;
 };
@@ -25,6 +27,7 @@ export class RateLimitRuleEntity {
   route: string;
   method: string;
   tier: number;
+  apiKey: string;
   options: RateLimitRuleOptions;
   createdAt: string;
 
@@ -33,6 +36,7 @@ export class RateLimitRuleEntity {
     this.route = params.route;
     this.method = params.method;
     this.tier = params.tier;
+    this.apiKey = params.api_key;
     this.options = params.options;
     this.createdAt = params.created_at;
   }